athena_avatar/main.py
#!/usr/bin/env python3
"""
Athena 3D Avatar Application
A cosmic AI companion with voice interaction and 3D visualization
Optimized for 12GB RAM with <250ms latency on mid-range devices
"""

import sys
import os
import logging
import asyncio
import gc
from pathlib import Path
from typing import Optional, Dict, Any

# PyQt6 imports
from PyQt6.QtWidgets import (QApplication, QMainWindow, QWidget, QVBoxLayout, 
                             QHBoxLayout, QPushButton, QLabel, QSlider, 
                             QCheckBox, QComboBox, QMessageBox, QProgressBar)
from PyQt6.QtCore import Qt, QTimer, QThread, pyqtSignal, QPropertyAnimation, QEasingCurve
from PyQt6.QtGui import QFont, QPalette, QColor, QPixmap, QIcon
from PyQt6.QtOpenGLWidgets import QOpenGLWidget

# PyTorch imports
import torch
import torch.nn as nn
import torch.optim as optim
from torch.utils.data import DataLoader

# 3D Graphics imports
import moderngl
import numpy as np
from OpenGL import GL

# Audio processing
import librosa
import soundfile as sf
import pydub

# Local imports
from core.memory_manager import MemoryManager
from core.model_optimizer import ModelOptimizer
from avatar.athena_model import AthenaModel
from avatar.voice_agent import BarkVoiceAgent, LipSyncAgent
from avatar.animation_controller import AnimationController
from rendering.neural_radiance_agent import NeuralRadianceAgent
from rendering.renderer_3d import Renderer3D
from ui.main_window import AthenaMainWindow
from utils.config_manager import ConfigManager
from utils.logger import setup_logger

class AthenaApp:
    """Main application class for Athena 3D Avatar"""
    
    def __init__(self):
        self.app = None
        self.main_window = None
        self.memory_manager = None
        self.model_optimizer = None
        self.athena_model = None
        self.voice_agent = None
        self.lip_sync_agent = None
        self.animation_controller = None
        self.nerf_agent = None
        self.renderer = None
        self.config = None
        self.logger = None
        
        # Performance tracking
        self.frame_times = []
        self.memory_usage = []
        self.latency_metrics = {}
        
    def initialize(self):
        """Initialize the application with all components"""
        try:
            # Setup logging
            self.logger = setup_logger()
            self.logger.info("Initializing Athena 3D Avatar Application")
            
            # Load configuration
            self.config = ConfigManager()
            self.config.load_config()
            
            # Initialize memory manager
            self.memory_manager = MemoryManager(max_ram_gb=12)
            self.memory_manager.initialize()
            
            # Initialize model optimizer
            self.model_optimizer = ModelOptimizer()
            self.model_optimizer.optimize_for_memory()
            
            # Initialize PyQt application
            self.app = QApplication(sys.argv)
            self.app.setApplicationName("Athena 3D Avatar")
            self.app.setApplicationVersion("1.0.0")
            
            # Set application style
            self.setup_application_style()
            
            # Initialize core components
            self.initialize_core_components()
            
            # Create main window
            self.main_window = AthenaMainWindow(self)
            self.main_window.show()
            
            self.logger.info("Application initialized successfully")
            return True
            
        except Exception as e:
            self.logger.error(f"Failed to initialize application: {e}")
            return False
    
    def setup_application_style(self):
        """Setup cosmic-themed application style"""
        # Create cosmic color palette
        cosmic_palette = QPalette()
        cosmic_palette.setColor(QPalette.ColorRole.Window, QColor(10, 10, 20))
        cosmic_palette.setColor(QPalette.ColorRole.WindowText, QColor(200, 200, 255))
        cosmic_palette.setColor(QPalette.ColorRole.Base, QColor(15, 15, 30))
        cosmic_palette.setColor(QPalette.ColorRole.AlternateBase, QColor(25, 25, 45))
        cosmic_palette.setColor(QPalette.ColorRole.ToolTipBase, QColor(20, 20, 35))
        cosmic_palette.setColor(QPalette.ColorRole.ToolTipText, QColor(200, 200, 255))
        cosmic_palette.setColor(QPalette.ColorRole.Text, QColor(220, 220, 255))
        cosmic_palette.setColor(QPalette.ColorRole.Button, QColor(30, 30, 50))
        cosmic_palette.setColor(QPalette.ColorRole.ButtonText, QColor(200, 200, 255))
        cosmic_palette.setColor(QPalette.ColorRole.Link, QColor(100, 150, 255))
        cosmic_palette.setColor(QPalette.ColorRole.Highlight, QColor(50, 100, 200))
        cosmic_palette.setColor(QPalette.ColorRole.HighlightedText, QColor(255, 255, 255))
        
        self.app.setPalette(cosmic_palette)
        
        # Set cosmic font
        cosmic_font = QFont("Segoe UI", 10)
        cosmic_font.setWeight(QFont.Weight.Medium)
        self.app.setFont(cosmic_font)
    
    def initialize_core_components(self):
        """Initialize all core components with memory optimization"""
        try:
            # Initialize Athena 3D model
            self.athena_model = AthenaModel()
            self.athena_model.load_optimized_model()
            
            # Initialize voice agents
            self.voice_agent = BarkVoiceAgent()
            self.lip_sync_agent = LipSyncAgent()
            
            # Initialize animation controller
            self.animation_controller = AnimationController()
            self.animation_controller.load_animations()
            
            # Initialize NeRF agent for advanced rendering
            self.nerf_agent = NeuralRadianceAgent()
            
            # Initialize 3D renderer
            self.renderer = Renderer3D()
            self.renderer.initialize()
            
            self.logger.info("Core components initialized successfully")
            
        except Exception as e:
            self.logger.error(f"Failed to initialize core components: {e}")
            raise
    
    def run(self):
        """Run the application main loop"""
        try:
            # Start performance monitoring
            self.start_performance_monitoring()
            
            # Run the application
            exit_code = self.app.exec()
            
            # Cleanup
            self.cleanup()
            
            return exit_code
            
        except Exception as e:
            self.logger.error(f"Application runtime error: {e}")
            return 1
    
    def start_performance_monitoring(self):
        """Start monitoring performance metrics"""
        # Monitor frame rate
        self.frame_timer = QTimer()
        self.frame_timer.timeout.connect(self.update_performance_metrics)
        self.frame_timer.start(1000)  # Update every second
    
    def update_performance_metrics(self):
        """Update performance metrics"""
        # Memory usage
        memory_usage = self.memory_manager.get_current_usage()
        self.memory_usage.append(memory_usage)
        
        # Frame rate calculation
        if hasattr(self, 'last_frame_time'):
            frame_time = self.get_current_time() - self.last_frame_time
            self.frame_times.append(frame_time)
            
            # Keep only last 60 frames for average calculation
            if len(self.frame_times) > 60:
                self.frame_times.pop(0)
        
        self.last_frame_time = self.get_current_time()
        
        # Log performance metrics
        if len(self.frame_times) > 0:
            avg_frame_time = sum(self.frame_times) / len(self.frame_times)
            fps = 1.0 / avg_frame_time if avg_frame_time > 0 else 0
            self.logger.info(f"Performance - FPS: {fps:.1f}, Memory: {memory_usage:.1f}GB")
    
    def get_current_time(self):
        """Get current time in seconds"""
        import time
        return time.time()
    
    def cleanup(self):
        """Cleanup resources"""
        try:
            self.logger.info("Cleaning up application resources")
            
            # Stop performance monitoring
            if hasattr(self, 'frame_timer'):
                self.frame_timer.stop()
            
            # Cleanup PyTorch models
            if self.athena_model:
                self.athena_model.cleanup()
            
            if self.voice_agent:
                self.voice_agent.cleanup()
            
            if self.lip_sync_agent:
                self.lip_sync_agent.cleanup()
            
            if self.nerf_agent:
                self.nerf_agent.cleanup()
            
            if self.renderer:
                self.renderer.cleanup()
            
            # Force garbage collection
            gc.collect()
            
            self.logger.info("Cleanup completed")
            
        except Exception as e:
            self.logger.error(f"Cleanup error: {e}")

def main():
    """Main entry point"""
    try:
        # Create and initialize application
        athena_app = AthenaApp()
        
        if not athena_app.initialize():
            print("Failed to initialize Athena application")
            return 1
        
        # Run the application
        return athena_app.run()
        
    except Exception as e:
        print(f"Critical error: {e}")
        return 1

if __name__ == "__main__":
    sys.exit(main())