athena_avatar/requirements.txt
# Athena 3D Avatar Application Requirements
# Core PyTorch and ML dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
diffusers>=0.21.0

# 3D Graphics and Rendering
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0
PyQt6-sip>=13.5.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6
moderngl>=5.8.0
trimesh>=3.20.0
pyglet>=2.0.0

# Voice and Audio Processing
torchaudio>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
webrtcvad>=2.0.10
pyttsx3>=2.90

# Neural Radiance Fields (NeRF)
nerfstudio>=0.3.0
pytorch3d>=0.7.0

# Image Processing and Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Numerical Computing
numpy>=1.24.0
scipy>=1.10.0
numba>=0.57.0

# Memory Management and Optimization
psutil>=5.9.0
memory-profiler>=0.60.0
line-profiler>=4.0.0

# Async and Concurrency
asyncio>=3.4.3
aiofiles>=23.0.0
aiohttp>=3.8.0

# Configuration and Settings
PyYAML>=6.0
configparser>=5.3.0

# Development and Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0

# Packaging and Distribution
pyinstaller>=5.13.0
setuptools>=68.0.0
wheel>=0.41.0

# Optional: GPU Acceleration
# torch-cuda>=2.0.0  # Uncomment for CUDA support
# cupy-cuda11x>=12.0.0  # Uncomment for CUDA support

# Optional: Advanced Features
# sentencepiece>=0.1.99  # For advanced text processing
# tokenizers>=0.13.0  # For tokenization